Cast to remove warning about implicit int conversion
Let's hope they know what they are doing ;-)

+++ Sources/Expat/xmlparse.c
@@ -3331,7 +3331,7 @@ storeAtts(XML_Parser parser, const ENCODING *enc, const char *attStr,
           */
           unsigned char step = 0;
           unsigned long mask = nsAttsSize - 1;
-          j = uriHash & mask; /* index into hash table */
+          j = (int)uriHash & mask; /* index into hash table */
           while (parser->m_nsAtts[j].version == version) {
             /* for speed we compare stored hash values first */
             if (uriHash == parser->m_nsAtts[j].hash) {
